Added efficient handlers for prepending and appending elements to the sorted list.
Imagine that you want to load a logfile where everything is likely (but not necessarily) in chronological order.
Appending to a sorted list now takes O(1) time instead of O(log(n)) time.
Prepending takes O(1) to find the spot... but still takes O(n) to shuffle up the list.  We'd have to change the underlying datastructure to fix that.